Extract admission permit into admission/permit.rs
