+++ crates/defra-agent/src/admission/mod.rs
@@ -10,6 +10,7 @@ pub(crate) use config::{backend_admission_configs_from_backends, BackendAdmissio
 pub(crate) use client::{
     scope_call, scope_request, AdmissionCallContext, AdmittedCompletionClient, CallKind,
 };
+pub(crate) use permit::AdmissionPermit;
 
 use std::collections::{HashMap, HashSet};
 use std::future::Future;
@@ -20,11 +21,9 @@ use std::sync::{Arc, Mutex};
 use anyhow::Result;
 use defra_node::EmbeddedNode;
 use rig::completion::{CompletionError, Usage};
-use tokio::sync::OwnedSemaphorePermit;
 
 use self::client::current_context;
 use self::controller::{BackendAdmissionController, InferenceCallRecord, PendingCallMetadata};
-use self::stream_guard::StreamGuardLifecycle;
 use crate::graphql::escape_graphql_string;
 
 #[derive(Clone)]
@@ -261,135 +260,6 @@ impl RegistryState {
                 BackendAdmissionController::new(pending.generation, pending.config, Arc::downgrade(registry)),
             );
         }
-    }
-}
-
-pub(crate) struct AdmissionPermit {
-    node: Arc<EmbeddedNode>,
-    controller: Arc<BackendAdmissionController>,
-    _permit: OwnedSemaphorePermit,
-    call: InferenceCallRecord,
-    _doc_id: String,
-    terminal: Option<PermitTerminal>,
-    finished: bool,
-}
-
-#[derive(Clone, Debug)]
-struct PermitTerminal {
-    call_state: &'static str,
-    failure_reason: Option<String>,
-    usage: Option<Usage>,
-}
-
-impl AdmissionPermit {
-    fn new(
-        node: Arc<EmbeddedNode>,
-        controller: Arc<BackendAdmissionController>,
-        permit: OwnedSemaphorePermit,
-        call: InferenceCallRecord,
-        doc_id: String,
-    ) -> Self {
-        Self {
-            node,
-            controller,
-            _permit: permit,
-            call,
-            _doc_id: doc_id,
-            terminal: None,
-            finished: false,
-        }
-    }
-
-    async fn finish_success(&mut self, usage: Option<Usage>) {
-        self.terminal = Some(PermitTerminal {
-            call_state: "completed",
-            failure_reason: None,
-            usage,
-        });
-        self.finish().await;
-    }
-
-    async fn finish_failure(&mut self, reason: &str) {
-        self.terminal = Some(PermitTerminal {
-            call_state: "failed",
-            failure_reason: Some(reason.to_string()),
-            usage: None,
-        });
-        self.finish().await;
-    }
-
-    async fn finish(&mut self) {
-        if self.finished {
-            return;
-        }
-        self.finished = true;
-        let terminal = self.terminal.clone().unwrap_or(PermitTerminal {
-            call_state: "completed",
-            failure_reason: None,
-            usage: None,
-        });
-        if let Err(error) = persist_existing_call_terminal(
-            self.node.clone(),
-            &self.call,
-            terminal.call_state,
-            terminal.failure_reason.as_deref(),
-            terminal.usage,
-        )
-        .await
-        {
-            tracing::warn!(call_id = %self.call.call_id, error = %error, "failed to persist terminal inference call state");
-        }
-    }
-}
-
-impl StreamGuardLifecycle for AdmissionPermit {
-    fn mark_stream_success(&mut self, usage: Option<Usage>) {
-        if self.terminal.is_none() {
-            self.terminal = Some(PermitTerminal {
-                call_state: "completed",
-                failure_reason: None,
-                usage,
-            });
-        }
-    }
-
-    fn mark_stream_error(&mut self, error: &CompletionError) {
-        self.terminal = Some(PermitTerminal {
-            call_state: "failed",
-            failure_reason: Some(error.to_string()),
-            usage: None,
-        });
-    }
-}
-
-impl Drop for AdmissionPermit {
-    fn drop(&mut self) {
-        self.controller.release_running();
-        if self.finished {
-            return;
-        }
-        self.finished = true;
-        let terminal = self.terminal.clone().unwrap_or(PermitTerminal {
-            call_state: "failed",
-            failure_reason: Some("StreamDroppedBeforeTerminalResponse".to_string()),
-            usage: None,
-        });
-        let node = self.node.clone();
-        let call_id = self.call.call_id.clone();
-        let call = self.call.clone();
-        spawn_persistence(async move {
-            if let Err(error) = persist_existing_call_terminal(
-                node,
-                &call,
-                terminal.call_state,
-                terminal.failure_reason.as_deref(),
-                terminal.usage,
-            )
-            .await
-            {
-                tracing::warn!(call_id = %call_id, error = %error, "failed to persist dropped inference call state");
-            }
-        });
     }
 }
